delete method -- minor issue

+++ bubbledocs-appserver/src/main/java/pt/tecnico/bubbledocs/domain/User.java
@@ -45,7 +45,8 @@ public class User extends User_Base {
 		
 		setBubbleDocs(null);
 		
-		this.getSession().delete();
+		if(this.getSession() != null)
+			this.getSession().delete();
 		setSession(null);
 
 		//Delete Object
